# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Zault API
  description: |
    Zault investment dashboard backend API.

    **API versioning:** All endpoints under `/api/**` except `/api/auth/login`, `/api/auth/register`, and `/api/health*` require the `X-API-Version: 1` request header. Missing or unsupported values return `400` with a structured error body.
  version: 0.0.1
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /api/tradebook/files:
    get:
      tags:
      - tradebook-controller
      operationId: listFiles
      parameters:
      - name: X-API-Version
        in: header
        description: API version — must be `1`
        required: true
        schema:
          type: string
          default: "1"
          enum:
          - "1"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TradeFileDto"
    post:
      tags:
      - tradebook-controller
      operationId: uploadFiles
      parameters:
      - name: files
        in: query
        required: true
        schema:
          type: array
          items:
            type: string
            format: binary
      - name: X-API-Version
        in: header
        description: API version — must be `1`
        required: true
        schema:
          type: string
          default: "1"
          enum:
          - "1"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/investments:
    get:
      tags:
      - investment-controller
      operationId: listInvestments
      parameters:
      - name: X-API-Version
        in: header
        description: API version — must be `1`
        required: true
        schema:
          type: string
          default: "1"
          enum:
          - "1"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InvestmentDto"
    post:
      tags:
      - investment-controller
      operationId: createInvestment
      parameters:
      - name: X-API-Version
        in: header
        description: API version — must be `1`
        required: true
        schema:
          type: string
          default: "1"
          enum:
          - "1"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvestmentRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/investments/{id}:
    delete:
      tags:
      - investment-controller
      operationId: deleteInvestment
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: X-API-Version
        in: header
        description: API version — must be `1`
        required: true
        schema:
          type: string
          default: "1"
          enum:
          - "1"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
    patch:
      tags:
      - investment-controller
      operationId: updateInvestmentAmount
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: X-API-Version
        in: header
        description: API version — must be `1`
        required: true
        schema:
          type: string
          default: "1"
          enum:
          - "1"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInvestmentAmountRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/tradebook/trades:
    get:
      tags:
      - tradebook-controller
      operationId: getTrades
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      - name: fileId
        in: query
        required: false
        schema:
          type: string
      - name: X-API-Version
        in: header
        description: API version — must be `1`
        required: true
        schema:
          type: string
          default: "1"
          enum:
          - "1"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TradesPageDto"
  /api/tradebook/trades/timeline:
    get:
      tags:
      - tradebook-controller
      operationId: getTradeTimeline
      parameters:
      - name: X-API-Version
        in: header
        description: API version — must be `1`
        required: true
        schema:
          type: string
          default: "1"
          enum:
          - "1"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TradeTimelineDto"
  /api/tradebook/allocations:
    get:
      tags:
      - tradebook-controller
      operationId: getAllocations
      parameters:
      - name: X-API-Version
        in: header
        description: API version — must be `1`
        required: true
        schema:
          type: string
          default: "1"
          enum:
          - "1"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AllocationsDto"
...
  /api/tradebook/files/{fileId}:
    delete:
      tags:
      - tradebook-controller
      operationId: deleteFile
      parameters:
      - name: fileId
        in: path
        required: true
        schema:
          type: string
      - name: X-API-Version
        in: header
        description: API version — must be `1`
        required: true
        schema:
          type: string
          default: "1"
          enum:
          - "1"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
components:
  schemas:
    CreateInvestmentRequest:
      type: object
      properties:
        category:
          type: string
        amount:
          type: number
...
    UpdateInvestmentAmountRequest:
      type: object
      properties:
        amount:
          type: number
    TradeDto:
      type: object
      properties:
        tradeId:
          type: string
        fileId:
          type: string
        symbol:
          type: string
        isin:
          type: string
        tradeDate:
          type: string
        exchange:
          type: string
        segment:
          type: string
        series:
          type: string
        tradeType:
          type: string
        auction:
          type: boolean
        quantity:
          type: number
        price:
          type: number
        orderId:
          type: string
        orderExecutionTime:
          type: string
    TradesPageDto:
      type: object
      properties:
        trades:
          type: array
          items:
            $ref: "#/components/schemas/TradeDto"
        totalCount:
          type: integer
          format: int64
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
    TradePoint:
      type: object
      properties:
        time:
          type: string
        tradeType:
          type: string
        amount:
          type: number
    TradeTimelineDto:
      type: object
      properties:
        trades:
          type: array
          items:
            $ref: "#/components/schemas/TradePoint"
    TradeFileDto:
      type: object
      properties:
        id:
          type: string
        filename:
          type: string
        rowCount:
          type: integer
          format: int32
        uploadedAt:
          type: string
    AllocationDto:
      type: object
      properties:
        isin:
          type: string
        symbol:
          type: string
        netQuantity:
          type: number
        investedAmount:
          type: number
    AllocationsDto:
      type: object
      properties:
        allocations:
          type: array
          items:
            $ref: "#/components/schemas/AllocationDto"
        totalInvested:
          type: number
    InvestmentDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        category:
          type: string
        amount:
          type: number
...